unify navbar components
