+++ src/pages/About.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import '../components/Project.css';
 
-import PageHeader from '../components/Header/PageHeader';
+import Header from '../components/Header/Header';
 import SideBar from '../components/SideBar/SideBar';
 import TopShortcut from '../components/TopShortcut/TopShortcut';
 import PageContent from './PageContainer';
@@ -15,7 +15,9 @@ export default function About() {
     
     return(
         <>
-        <PageHeader />
+        <Header 
+        pageType='nav-page animated fadeIn' 
+        aboutType='nav-desktop-page link' />
         <PageContent>
             <div className='project animated fadeIn'>
                 <div className='project_body'>
