+++ src/pages/FASSB.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import '../components/Project.css';
 
-import PageHeader from '../components/Header/PageHeader';
+import Header from '../components/Header/Header';
 import SideBar from '../components/SideBar/SideBar';
 import TopShortcut from '../components/TopShortcut/TopShortcut';
 import PageContent from './PageContainer';
@@ -25,7 +25,9 @@ export default function FASSB() {
     
     return(
         <>
-        <PageHeader />
+        <Header 
+        pageType='nav-page animated fadeIn' 
+        aboutType='nav-desktop-page link' />
         <PageContent>
             <div className='project animated fadeIn'>
                 <img className='project_banner' src={Banner} alt='fassb' />
