+++ src/pages/Home.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import '../components/Home.css';
 
-import HomeHeader from '../components/Header/HomeHeader';
+import Header from '../components/Header/Header';
 import Carousel from '../components/Carousel/Carousel';
 import SideBar from '../components/SideBar/SideBar';
 import BlueBlock from '../components/BlueBlock/BlueBlock';
@@ -12,7 +12,9 @@ export default function Home() {
     
     return (
         <>
-        <HomeHeader />
+        <Header 
+        pageType="nav-home animated fadeIn" 
+        aboutType="nav-desktop-home link" />
         <div className='intro_sm animated fadeIn'>
             <h1><strong>DESIGNER</strong></h1>
             <h6><strong>ENGINEERING STUDENT</strong></h6>
